VS2005
git-svn-id: https://svn.code.sf.net/p/ultravnc/code/UltraVNC%20Project%20Root/UltraVNC@21 0bd26501-c230-0410-ab9e-a7946180fe3b

+++ rdr/FdInStream.cxx
@@ -186,7 +186,7 @@ static void gettimeofday(struct timeval* tv, void*)
   } else {
     struct timeb tb;
     ftime(&tb);
-    tv->tv_sec = tb.time;
+    tv->tv_sec = (long)tb.time;
     tv->tv_usec = tb.millitm * 1000;
   }
 }
